patter devision into logical parts
