+++ js/stats-template.js
@@ -1,0 +1,15 @@
+const statsTemplate = `
+  <ul class="stats">
+    <li class="stats__result stats__result--wrong"></li>
+    <li class="stats__result stats__result--slow"></li>
+    <li class="stats__result stats__result--fast"></li>
+    <li class="stats__result stats__result--correct"></li>
+    <li class="stats__result stats__result--unknown"></li>
+    <li class="stats__result stats__result--unknown"></li>
+    <li class="stats__result stats__result--unknown"></li>
+    <li class="stats__result stats__result--unknown"></li>
+    <li class="stats__result stats__result--unknown"></li>
+    <li class="stats__result stats__result--unknown"></li>
+  </ul>`;
+
+export default statsTemplate;
